+++ R/derive_var_ady.R
@@ -14,11 +14,15 @@
 #'
 #'   A date or date-time object column is expected.
 #'
+#'   Refer to `derive_var_dt()` to impute and derive a date from a date character vector to a date object.
+#'
 #'   The default is `TRTSDT`.
 #'
 #' @param date The end date column for which the study day should be derived
 #'
 #'   A date or date-time object column is expected.
+#'
+#'   Refer to `derive_var_dt()` to impute and derive a date from a date character vector to a date object.
 #'
 #'   The default is `ADT`
 #'